WIP: Cancel requests tests
